optimize the code

+++ src/main/kotlin/com/example/petheaven/controller/EquipmentController.kt
@@ -1,41 +1,75 @@
 package com.example.petheaven.controller
 
-import com.example.petheaven.model.Equipment
+import com.example.petheaven.model.Result
 import com.example.petheaven.service.EquipmentService
+import com.example.petheaven.service.UserService
+import com.example.petheaven.vo.EquipmentVo
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 
 @RestController
-@RequestMapping("/api/shop")
-class EquipmentController(val service: EquipmentService) {
-    @GetMapping
+@RequestMapping("/equipment")
+class EquipmentController(val service: EquipmentService, val userService: UserService) {
+    @GetMapping("/shop")
     fun index(): ResponseEntity<Any> {
-        return ResponseEntity.ok(service.getShopList())
-    }
-
-    @GetMapping("/equipment")
-    fun backstage(): ResponseEntity<Any> {
-        return ResponseEntity.ok(service.getEquipmentList())
+        return try {
+            ResponseEntity.ok(service.getShopList())
+        } catch (exception: Exception) {
+            ResponseEntity.ok(Result.errorResult())
+        }
     }
 
     @GetMapping("/{id}")
     fun info(@PathVariable id: Long): ResponseEntity<Any> {
-        return ResponseEntity.ok(service.getInfo(id))
+        return try {
+            ResponseEntity.ok(service.getInfo(id))
+        } catch (exception: Exception) {
+            ResponseEntity.ok(Result.errorResult())
+        }
     }
 
-    @PostMapping("/create")
-    fun create(@RequestBody equipment: Equipment): ResponseEntity<Any> {
-        service.create(equipment)
-        return ResponseEntity.ok(service.create(equipment))
+    @GetMapping("/admin/{userId}")
+    fun backstage(@PathVariable userId: Long): ResponseEntity<Any> {
+        return try {
+            checkIsAdmin(userId)
+            ResponseEntity.ok(service.getEquipmentList())
+        } catch (exception: Exception) {
+            ResponseEntity.ok(Result.errorResult())
+        }
     }
 
-    @PutMapping("/edit")
-    fun edit(@RequestBody equipment: Equipment): Any {
-        return ResponseEntity.ok(service.edit(equipment))
+    @PostMapping("/admin/create")
+    fun create(@RequestBody equipmentVo: EquipmentVo): ResponseEntity<Any> {
+        return try {
+            checkIsAdmin(equipmentVo.userId)
+            ResponseEntity.ok(service.create(equipmentVo))
+        } catch (exception: Exception) {
+            ResponseEntity.ok(Result.errorResult())
+        }
     }
 
-    @DeleteMapping("/delete/{id}")
-    fun delete(@PathVariable id: Long): Any {
-        return ResponseEntity.ok(service.delete(id))
+    @PutMapping("/admin/edit/{id}")
+    fun edit(@PathVariable id: Long, @RequestBody equipmentVo: EquipmentVo): Any {
+        return try {
+            checkIsAdmin(equipmentVo.userId)
+            ResponseEntity.ok(service.edit(id, equipmentVo))
+        } catch (exception: Exception) {
+            ResponseEntity.ok(Result.errorResult())
+        }
+    }
+
+    @DeleteMapping("/admin/delete/{userId}/{id}")
+    fun delete(@PathVariable userId: Long, @PathVariable id: Long): Any {
+        return try {
+            checkIsAdmin(userId)
+            ResponseEntity.ok(service.delete(id))
+        } catch (exception: Exception) {
+            ResponseEntity.ok(Result.errorResult())
+        }
+    }
+
+    fun checkIsAdmin(userId : Long) {
+        val isAdmin = userService.checkIsAdmin(userId).message
+        if(isAdmin != true) throw Exception()
     }
 }
